requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
kucoin-python==1.0.11
requests==2.28.1
urllib3==1.26.12
websockets==10.4
